feat(prophecy/unit): Add parameters class validator
