+++ src/army-list/army-list.dto.ts
@@ -1,6 +1,8 @@
 import { ArmyList } from "./army-list.entity";
 import { ArmyListUnitDTO } from "./army-list-unit/army-list-unit.dto";
-import { ArmyListUnitCredentialsDTO } from "@army-list/army-list-unit/army-list-unit-credentials.dto";
+import {
+    ArmyListUnitCredentialsDTO
+} from "@army-list/army-list-unit/army-list-unit-credentials.dto";
 import { IsArray, IsBoolean, IsDefined, IsNotEmpty, IsNumber, IsString, Min, ValidateNested } from "class-validator";
 import { Type } from "class-transformer";
 import { ArmyListUnit } from "@army-list/army-list-unit/army-list-unit.entity";
