+++ tests/api/army-list/get.spec.ts
@@ -102,5 +102,4 @@ describe("armies-lists/get", () => {
 
         expect(res.status).toEqual(HttpStatus.NOT_FOUND)
     });
-
 });
